login with auth system

+++ src/utils/manageData.ts
@@ -2,10 +2,13 @@ const PORT = 4200;
 
 export const getData = async (page: number = 1) => {
   try {
+    const token = localStorage.getItem("token");
+    if (!token) return;
     const response = await fetch(`http://localhost:${PORT}/data?page=${page}`, {
       method: "GET",
       headers: {
         "Content-Type": "application/json",
+        "x-auth-token": token,
       },
     });
     const data = await response.json();
@@ -22,10 +25,14 @@ export const postData = async (data: any) => {
   for (let i = 0; i < data.length; i += batchSize) {
     const batch = data.slice(i, i + batchSize);
     try {
+      // verify if the token exists
+      const token = localStorage.getItem("token");
+      if (!token) return;
       const response = await fetch(`http://localhost:${PORT}/data`, {
         method: "POST",
         headers: {
           "Content-Type": "application/json",
+          "x-auth-token": token,
         },
         body: JSON.stringify(batch),
       });
@@ -34,7 +41,6 @@ export const postData = async (data: any) => {
         console.log(response.status, response.text);
         throw new Error("Request failed!");
       }
-
       const json = await response.json();
       console.log(json);
     } catch (err) {
@@ -45,8 +51,13 @@ export const postData = async (data: any) => {
 
 export const deleteData = async () => {
   try {
+    const token = localStorage.getItem("token");
+    if (!token) return;
     const response = await fetch(`http://localhost:${PORT}/data`, {
       method: "DELETE",
+      headers: {
+        "x-auth-token": token,
+      },
     });
     console.log(response, response.status);
   } catch (err) {
